+++ tests/test_ugrid1d.py
@@ -610,3 +610,20 @@ def test_ugrid1d_create_data_array():
     # Error on size
     with pytest.raises(ValueError, match="Conflicting sizes"):
         grid.create_data_array([1, 2, 3, 4], facet="node")
+
+
+def test_ugrid1d_format_connectivity():
+    grid = grid1d()
+    assert isinstance(grid.edge_node_connectivity, np.ndarray)
+    assert isinstance(
+        grid.format_connectivity_as_sparse(grid.edge_node_connectivity),
+        sparse.csr_matrix,
+    )
+    assert isinstance(grid.node_node_connectivity, sparse.csr_matrix)
+    assert isinstance(
+        grid.format_connectivity_as_dense(grid.node_node_connectivity), np.ndarray
+    )
+    assert isinstance(
+        grid.format_connectivity_as_sparse(grid.node_node_connectivity.tocoo()),
+        sparse.csr_matrix,
+    )