+++ docs/changelog.rst
@@ -17,6 +17,17 @@ Changed
   :meth:`xugrid.Ugrid1d.intersect_linestring` now actually conduct spatial
   indexing instead of just returning the object.
 
+Added
+~~~~~
+
+- Added :meth:`xugrid.Ugrid1d.format_connectivity_as_dense`,
+  :meth:`xugrid.Ugrid1d.format_connectivity_as_sparse`,
+  :meth:`xugrid.Ugrid2d.format_connectivity_as_dense`, and
+  :meth:`xugrid.Ugrid2d.format_connectivity_as_sparse` utilities to convert
+  between sparse and dense rectangular form of the connectivity arrays.
+- Added :attr:`xugrid.Ugrid1d.edge_edge_connectivity` and
+  :attr:`xugrid.Ugrid2d.edge_edge_connectivity`.
+
 [0.12.4] 2025-03-05
 -------------------